backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
typing
pydantic[email]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosqlite
jose
